+++ packages/plugins/rsbuild-plugin-extension-hmr/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @addfox/rsbuild-plugin-extension-hmr
 
+## 0.2.2
+
+### Patch Changes
+
+- feat: add create arg
+- Updated dependencies
+  - @addfox/launcher@0.2.2
+  - @addfox/common@0.2.2
+  - @addfox/core@0.2.2
+
 ## 0.2.1
 
 ### Patch Changes